refactor(pagination): tweak styles, move examples, fix docs
