+++ projects/cashmere-examples/src/lib/pagination-load-more/pagination-load-more-example.css
@@ -1,4 +1,3 @@
-.hc-form-field {
-    margin: 0 25px 15px 0;
-    display: inline-flex;
+table {
+    width: 100%;
 }
