+++ projects/cashmere-examples/src/lib/pagination-standard/pagination-standard-example.css
@@ -1,4 +1,3 @@
-.hc-form-field {
-    margin: 0 25px 15px 0;
-    display: inline-flex;
+table {
+    width: 100%;
 }
