+++ projects/cashmere-examples/src/lib/pagination-standard/pagination-standard-example.html
@@ -1,14 +1,37 @@
-<hc-form-field inline>
-    <hc-label>Number of items</hc-label>
-    <input hcInput type="number" [(ngModel)]="length">
-</hc-form-field>
-<hc-form-field inline>
-    <hc-label>Current page number:</hc-label>
-    <input hcInput type="number" [(ngModel)]="currentPage">
-</hc-form-field>
-<br>
-<hc-checkbox id="hide-page-size" [(ngModel)]="hidePageSize" [checked]="hidePageSize">Hide page size selector</hc-checkbox>
+<table hc-table [dataSource]="dataSource">
 
-<hr>
+    <!-- Position Column -->
+    <ng-container hcColumnDef="position">
+        <th hc-header-cell *hcHeaderCellDef>No.</th>
+        <td hc-cell *hcCellDef="let element">{{element.position}}</td>
+    </ng-container>
 
-<hc-pagination [length]="length" [pageNumber]="currentPage" [hidePageSize]="hidePageSize"></hc-pagination>
+    <!-- Name Column -->
+    <ng-container hcColumnDef="name">
+        <th hc-header-cell *hcHeaderCellDef>Name</th>
+        <td hc-cell *hcCellDef="let element">{{element.name}}</td>
+    </ng-container>
+
+    <!-- Weight Column -->
+    <ng-container hcColumnDef="weight">
+        <th hc-header-cell *hcHeaderCellDef>Weight</th>
+        <td hc-cell *hcCellDef="let element">{{element.weight}}</td>
+    </ng-container>
+
+    <!-- Symbol Column -->
+    <ng-container hcColumnDef="symbol">
+        <th hc-header-cell *hcHeaderCellDef>Symbol</th>
+        <td hc-cell *hcCellDef="let element">{{element.symbol}}</td>
+    </ng-container>
+
+    <tr hc-header-row *hcHeaderRowDef="displayedColumns"></tr>
+    <tr hc-row *hcRowDef="let row; columns: displayedColumns;"></tr>
+</table>
+
+<hc-pagination
+    [length]="length"
+    [(pageNumber)]="pageNumber"
+    pageSize="10"
+    [hidePageSize]="false"
+    [pageSizeOptions]="pageOpts">
+</hc-pagination>
