--- projects/cashmere-examples/src/lib/table-load-more/table-load-more-example.css
@@ -1,3 +1,0 @@
-table {
-    width: 100%;
-}
