--- projects/cashmere-examples/src/lib/table-load-more/table-load-more-example.html
@@ -1,31 +1,0 @@
-<table hc-table [dataSource]="dataSource">
-
-    <!-- Position Column -->
-    <ng-container hcColumnDef="position">
-        <th hc-header-cell *hcHeaderCellDef>No.</th>
-        <td hc-cell *hcCellDef="let element">{{element.position}}</td>
-    </ng-container>
-
-    <!-- Name Column -->
-    <ng-container hcColumnDef="name">
-        <th hc-header-cell *hcHeaderCellDef>Name</th>
-        <td hc-cell *hcCellDef="let element">{{element.name}}</td>
-    </ng-container>
-
-    <!-- Weight Column -->
-    <ng-container hcColumnDef="weight">
-        <th hc-header-cell *hcHeaderCellDef>Weight</th>
-        <td hc-cell *hcCellDef="let element">{{element.weight}}</td>
-    </ng-container>
-
-    <!-- Symbol Column -->
-    <ng-container hcColumnDef="symbol">
-        <th hc-header-cell *hcHeaderCellDef>Symbol</th>
-        <td hc-cell *hcCellDef="let element">{{element.symbol}}</td>
-    </ng-container>
-
-    <tr hc-header-row *hcHeaderRowDef="displayedColumns"></tr>
-    <tr hc-row *hcRowDef="let row; columns: displayedColumns;"></tr>
-</table>
-
-<hc-load-more-pagination [length]="length" pageSize="5"></hc-load-more-pagination>
