--- projects/cashmere-examples/src/lib/table-page/table-page-example.css
@@ -1,3 +1,0 @@
-table {
-    width: 100%;
-}
