+++ projects/cashmere/src/lib/pagination/index.ts
@@ -2,3 +2,4 @@ export {PaginationComponent} from './pagination.component';
 export {LoadMorePaginationComponent} from './load-more-pagination.component';
 export {PageEvent} from './page-event';
 export {PaginationModule} from './pagination.module';
+export {BasePaginationComponent} from './base-pagination';
