+++ projects/cashmere/src/lib/pagination/load-more-pagination.component.scss
@@ -1,0 +1,5 @@
+.hc-button {
+    display: block;
+    min-width: 30%;
+    margin: 20px auto;
+}
