+++ projects/cashmere/src/lib/pagination/pagination.component.scss
@@ -7,6 +7,7 @@
     justify-content: space-between;
     align-items: center;
     flex-wrap: wrap;
+    margin-top: 15px;
 }
 
 .hc-page-detail-container {
