+++ projects/cashmere/src/lib/pagination/pagination.component.ts
@@ -4,6 +4,7 @@ import {BasePaginationComponent} from './base-pagination';
 
 /** The pagination control enables the user to navigate across paged content.
  * Although commonly used with tables and data grids, this control may be used any place where paged data is used.
+ * @inheritdoc
  * */
 @Component({
     selector: 'hc-pagination',
@@ -11,7 +12,7 @@ import {BasePaginationComponent} from './base-pagination';
     styleUrls: ['./pagination.component.scss']
 })
 export class PaginationComponent extends BasePaginationComponent implements OnInit {
-    /** The set of provided page size options to display to the user. */
+    /** The set of provided page size options to display to the user. *Defaults to [10, 20, 50].* */
     @Input()
     get pageSizeOptions(): number[] {
         return this._pageSizeOptions;
