+++ projects/cashmere/src/lib/pagination/pagination.md
@@ -1,6 +1,6 @@
-The pagination control enables the user to navigate across paged content. Although commonly used with tables and data grids, this control may be used any place where paged data is used.
+The pagination controls enable the user to navigate across paged content. Although commonly used with tables and data grids, this control may be used any place where paged data is used.
 
-##### Responsiveness
+##### Responsiveness for Standard Pagination Control
 
 Normally, the pagination control displays a previous button, a next button, and nine page options (some of which may be ellipses). For example, if on page 13/28, the options would be as follows:
