+++ projects/cashmere/src/lib/pagination/pagination.module.ts
@@ -5,10 +5,11 @@ import {SelectModule} from '../select/select.module';
 import {IconModule} from '../icon/icon.module';
 import {PaginationComponent} from './pagination.component';
 import {LoadMorePaginationComponent} from './load-more-pagination.component';
+import {BasePaginationComponent} from './base-pagination';
 
 @NgModule({
     imports: [CommonModule, ButtonModule, IconModule, SelectModule],
-    declarations: [PaginationComponent, LoadMorePaginationComponent],
-    exports: [PaginationComponent, LoadMorePaginationComponent]
+    declarations: [PaginationComponent, LoadMorePaginationComponent, BasePaginationComponent],
+    exports: [PaginationComponent, LoadMorePaginationComponent, BasePaginationComponent]
 })
 export class PaginationModule {}
