+++ projects/cashmere/src/lib/sass/tables.scss
@@ -249,13 +249,3 @@ table.hc-table {
         }
     }
 }
-
-.hc-table + hc-pagination {
-    margin-top: 15px;
-    display: block;
-}
-
-.hc-table + hc-load-more-pagination {
-    margin-top: 20px;
-    display: block;
-}
