+++ projects/cashmere/src/lib/table/table.md
@@ -164,6 +164,12 @@ it is contained in the reduced string, and the row would be displayed in the tab
 To override the default filtering behavior, a custom `filterPredicate` function can be set which
 takes a data object and filter string and returns true if the data object is considered a match.
 
+#### Pagination
+
+There are cashmere pagination components available for usage. To learn about how to implement those, visit the
+[pagination component examples](/components/pagination/examples).
+
+
 #### Selection
 
 Right now there is no formal support for adding a selection UI to the table, but Cashmere
